big logical refactor
